Add client flow test coverage
